// frontend/src/api/generated/models/generateReviewedQuestionsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { GenerateReviewedQuestionsResponseDtoMeta } from "./generateReviewedQuestionsResponseDtoMeta";
import type { ReviewedQuestionsDataResponseDto } from "./reviewedQuestionsDataResponseDto";

export interface GenerateReviewedQuestionsResponseDto {
  statusCode: number;
  message: string;
  count: number;
  meta: GenerateReviewedQuestionsResponseDtoMeta;
  data: ReviewedQuestionsDataResponseDto;
}